separate random agent into its own file
