+++ ubikagent/agent/__init__.py
@@ -1,4 +1,5 @@
-from .agent import Agent, RandomGymAgent, RandomAgent
+from .agent import Agent
+from .random import RandomGymAgent, RandomAgent
 from .dqn import DQNAgent
 from .ddpg import DDPGAgent
 from .sarsa import SarsaAgent
